Allow hostname specification via compiler commandline
The ota platformio environment is nice, but it defaults
to using the "openevse.local" hostname/port for uploads;
in an environment which has a production OpenEVSE wifi
instance, this may not be desirable.

Allow specification of the hostname during the build
via a commandline option, and if built under the
openevse_ota environment, set that name to the value set
in the ESP_WIFI_HOSTNAME environment variable.
Upload OTA to that target as well.